+++ src/events/CampaignEvent.php
@@ -8,4 +8,10 @@ use yii\base\Event;
 class CampaignEvent extends Event
 {
     public ?Campaign $campaign = null;
+
+    /**
+     * @var bool Whether the campaign send is valid and should proceed. Set to
+     * `false` in a `beforeSend` handler to cancel the send.
+     */
+    public bool $isValid = true;
 }